dbeaver/pro#8999 fixes
