+++ plugins/org.jkiss.dbeaver.registry/OSGI-INF/l10n/bundle.properties
@@ -14,6 +14,7 @@ extension-point.org.jkiss.dbeaver.fileSystem.name = Virtual file systems
 extension-point.org.jkiss.dbeaver.language.name = Languages
 extension-point.org.jkiss.dbeaver.task.name = Tasks and task types
 extension-point.org.jkiss.dbeaver.productFeature.name = Product feature
+extension-point.org.jkiss.dbeaver.nativeClient.name=Native client
 
 driver.category.sql.name = SQL
 driver.category.sql.description = SQL (relational) databases
