+++ plugins/org.jkiss.dbeaver.registry/schema/org.jkiss.dbeaver.nativeClient.exsd
@@ -1,0 +1,12 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<schema targetNamespace="org.jkiss.dbeaver.core" xmlns="http://www.w3.org/2001/XMLSchema">
+<annotation>
+    <annotation>
+        <appinfo>
+            <meta.schema plugin="org.jkiss.dbeaver.registry" id="org.jkiss.dbeaver.nativeClient" name="Native client"/>
+        </appinfo>
+
+        <documentation>
+        </documentation>
+
+</schema>
