+++ plugins/org.jkiss.dbeaver.registry/src/org/jkiss/dbeaver/registry/nativeclient/NativeClientFileDescriptor.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.jkiss.dbeaver.registry.tool;
+package org.jkiss.dbeaver.registry.nativeclient;
 
 import org.eclipse.core.runtime.IConfigurationElement;
 import org.jkiss.dbeaver.model.impl.AbstractDescriptor;
